+++ src/app.html
@@ -1,12 +1,12 @@
 <!doctype html>
-<html lang="en" data-theme="cyberpunk">
+<html lang="en" class="dark">
 	<head>
 		<meta charset="utf-8" />
 		<link rel="icon" href="%sveltekit.assets%/favicon.png" />
 		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		%sveltekit.head%
 	</head>
-	<body data-sveltekit-preload-data="hover"  style="width: 400px; height: 500px; max-width: 400px">
+	<body data-sveltekit-preload-data="hover" data-theme="wintry" style="width: 400px; height: 500px; max-width: 400px">
 		<div style="display: contents">%sveltekit.body%</div>
 	</body>
 </html>
