+++ src/lib/components/AuthorizationNew.svelte
@@ -1,56 +1,35 @@
+<script>
+	import Icon from '@iconify/svelte';
+</script>
+
 <div
-  class="justify-center items-stretch bg-black bg-opacity-30 flex w-full flex-col mt-3 p-4 rounded-2xl"
+	class="justify-center items-stretch bg-black bg-opacity-50 flex w-full flex-col mt-3 p-4 rounded-2xl"
 >
-  <div
-    class="text-white text-opacity-70 text-xs font-semibold leading-4 tracking-[2.4000000000000004px]"
-  >
-    AUTHORISATION
-  </div>
-  <div
-    class="text-white text-2xl font-semibold leading-7 whitespace-nowrap mt-2"
-  >
-    coracle.social
-  </div>
-  <div class="text-white text-base leading-5 whitespace-nowrap mt-2">
-    would like to:
-  </div>
-  <div class="items-center flex justify-between gap-3 mt-2">
-    <img
-      loading="lazy"
-      alt=""
-      src="https://cdn.builder.io/api/v1/image/assets/TEMP/67e2e411-1462-4d78-ac7d-8c0902f9e0a5?apiKey=f7abd25632984a63bf82dda361df2c86&"
-      class="aspect-[1.5] object-contain object-center w-3 stroke-[1.4px] stroke-teal-400 overflow-hidden shrink-0 max-w-full my-auto"
-    />
-    <div
-      class="text-white text-base leading-5 self-stretch grow whitespace-nowrap"
-    >
-      read your public key
-    </div>
-  </div>
-  <div class="items-center flex justify-between gap-3 mt-2">
-    <img
-      loading="lazy"
-      alt=""
-      src="https://cdn.builder.io/api/v1/image/assets/TEMP/5267675d-fc62-4d2e-8c91-327f6dcf960a?apiKey=f7abd25632984a63bf82dda361df2c86&"
-      class="aspect-[1.5] object-contain object-center w-3 stroke-[1.4px] stroke-teal-400 overflow-hidden shrink-0 max-w-full my-auto"
-    />
-    <div
-      class="text-white text-base leading-5 self-stretch grow whitespace-nowrap"
-    >
-      sign some events
-    </div>
-  </div>
-  <div class="items-center flex justify-between gap-3 mt-2">
-    <img
-      loading="lazy"
-      alt=""
-      src="https://cdn.builder.io/api/v1/image/assets/TEMP/53ed6ada-d122-4fee-b9b0-4e41ddf830e2?apiKey=f7abd25632984a63bf82dda361df2c86&"
-      class="aspect-[1.5] object-contain object-center w-3 stroke-[1.4px] stroke-teal-400 overflow-hidden shrink-0 max-w-full my-auto"
-    />
-    <div
-      class="text-white text-base leading-5 self-stretch grow whitespace-nowrap"
-    >
-      do some other stuff
-    </div>
-  </div>
+	<div
+		class="text-white text-opacity-70 text-xs font-semibold leading-4 tracking-[2.4000000000000004px]"
+	>
+		AUTHORISATION
+	</div>
+	<div class="text-white text-2xl font-semibold leading-7 whitespace-nowrap mt-2">
+		coracle.social
+	</div>
+	<div class="text-white text-base leading-5 whitespace-nowrap mt-2">would like to:</div>
+	<div class="items-center flex justify-between gap-3 mt-2">
+		<Icon icon="mdi:check" width={16} class="text-teal-400" />
+		<div class="text-white text-base leading-5 self-stretch grow whitespace-nowrap">
+			read your public key
+		</div>
+	</div>
+	<div class="items-center flex justify-between gap-3 mt-2">
+		<Icon icon="mdi:check" width={16} class="text-teal-400" />
+		<div class="text-white text-base leading-5 self-stretch grow whitespace-nowrap">
+			sign some events
+		</div>
+	</div>
+	<div class="items-center flex justify-between gap-3 mt-2">
+		<Icon icon="mdi:check" width={16} class="text-teal-400" />
+		<div class="text-white text-base leading-5 self-stretch grow whitespace-nowrap">
+			do some other stuff
+		</div>
+	</div>
 </div>
