+++ src/lib/components/AuthorizedApp.svelte
@@ -1,9 +1,11 @@
 <script lang="ts">
+	import Icon from '@iconify/svelte';
+
 	let daysLeft = 2;
 </script>
 
 <div
-	class="justify-center items-stretch bg-black bg-opacity-30 flex w-full flex-col mt-3 p-4 rounded-2xl"
+	class="justify-center items-stretch bg-black bg-opacity-50 flex w-full flex-col mt-3 p-4 rounded-2xl"
 >
 	<div class="justify-between items-stretch flex gap-0">
 		<div
@@ -17,11 +19,8 @@
 	</div>
 	<div class="justify-between items-stretch flex gap-5 mt-2">
 		<div class="text-white text-2xl font-semibold leading-7">coracle.social</div>
-		<img
-			loading="lazy"
-			alt=""
-			src="https://cdn.builder.io/api/v1/image/assets/TEMP/56345f55-265f-4f07-ae46-d38f07e06956?apiKey=f7abd25632984a63bf82dda361df2c86&"
-			class="aspect-square object-contain object-center w-4 stroke-[1.4px] stroke-white stroke-opacity-30 overflow-hidden self-center shrink-0 max-w-full my-auto"
-		/>
+		<button class="btn btn-sm btn-ghost px-0 py-0">
+			<Icon icon="mingcute:settings-2-line" width={26} class="opacity-40" />
+		</button>
 	</div>
 </div>
