+++ src/lib/components/RecentActivity.svelte
@@ -2,51 +2,41 @@
 </script>
 
 <div
-  class="justify-center items-stretch bg-black bg-opacity-30 flex w-full flex-col mt-3 p-4 rounded-2xl"
+	class="justify-center items-stretch bg-black bg-opacity-50 flex w-full flex-col mt-3 p-4 rounded-2xl"
 >
-  <div
-    class="text-white text-opacity-70 text-xs font-semibold leading-4 tracking-[2.4000000000000004px]"
-  >
-    RECENT ACTIVITY
-  </div>
-  <div class="justify-between items-stretch flex gap-5 mt-2">
-    <div class="text-white text-base leading-5">Note</div>
-    <div
-      class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto"
-    >
-      A hour ago
-    </div>
-  </div>
-  <div class="justify-between items-stretch flex gap-5 mt-2">
-    <div class="text-white text-base leading-5">Reply</div>
-    <div
-      class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto"
-    >
-      A hour ago
-    </div>
-  </div>
-  <div class="justify-between items-stretch flex gap-5 mt-2">
-    <div class="text-white text-base leading-5">Reaction</div>
-    <div
-      class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto"
-    >
-      A hour ago
-    </div>
-  </div>
-  <div class="justify-between items-stretch flex gap-5 mt-2">
-    <div class="text-white text-base leading-5">Note</div>
-    <div
-      class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto"
-    >
-      A hour ago
-    </div>
-  </div>
-  <div class="justify-between items-stretch flex gap-5 mt-2">
-    <div class="text-white text-base leading-5">Highlight</div>
-    <div
-      class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto"
-    >
-      A hour ago
-    </div>
-  </div>
+	<div
+		class="text-white text-opacity-70 text-xs font-semibold leading-4 tracking-[2.4000000000000004px]"
+	>
+		RECENT ACTIVITY
+	</div>
+	<div class="justify-between items-stretch flex gap-5 mt-2">
+		<div class="text-white text-base leading-5">Note</div>
+		<div class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto">
+			A hour ago
+		</div>
+	</div>
+	<div class="justify-between items-stretch flex gap-5 mt-2">
+		<div class="text-white text-base leading-5">Reply</div>
+		<div class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto">
+			A hour ago
+		</div>
+	</div>
+	<div class="justify-between items-stretch flex gap-5 mt-2">
+		<div class="text-white text-base leading-5">Reaction</div>
+		<div class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto">
+			A hour ago
+		</div>
+	</div>
+	<div class="justify-between items-stretch flex gap-5 mt-2">
+		<div class="text-white text-base leading-5">Note</div>
+		<div class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto">
+			A hour ago
+		</div>
+	</div>
+	<div class="justify-between items-stretch flex gap-5 mt-2">
+		<div class="text-white text-base leading-5">Highlight</div>
+		<div class="text-white text-opacity-30 text-xs leading-4 self-center whitespace-nowrap my-auto">
+			A hour ago
+		</div>
+	</div>
 </div>
