+++ src/routes/+layout.svelte
@@ -2,8 +2,8 @@
 	import './styles.css';
 </script>
 
-<div class="app flex flex-col h-screen">
-	<main class="flex-grow">
+<div class="app flex flex-col h-screen w-full">
+	<main class="flex-grow w-full">
 		<slot />
 	</main>
 </div>
